feat: Extract common methods from BuildWorker
